requirements.txt
django==1.4.5
django-registration==0.8
django-tinymce==1.5
gevent==0.13.6
greenlet==0.3.1
gunicorn==0.14.5
ipaddr==2.1.10
kombu==2.1.8
lxml==3.4.2
mailer==0.7
ncclient==0.4.3
paramiko==1.7.7.1
pycrypto==2.6
pyparsing==1.5.6
python-dateutil==1.5
python-memcached==1.48
djangorestframework==2.3.14
